Initial Commit
We generate PSLG instances by first generating a random number of rectangles within a large rectangle. The PSLG considers the corners of these rectangles as well as the intersection points as vertices. We then decompose this PSLG into rectangles by shooting axes-parallel rays from reflex vertices. This decomposition is suboptimal but within a 2-factor of an optimal.